more work on fluid and energy handlers, changed back to default handlers for basic multi machine
