+++ src/main/java/muramasa/antimatter/tile/multi/TileEntityBasicMultiMachine.java
@@ -16,8 +16,8 @@ public class TileEntityBasicMultiMachine extends TileEntityMultiMachine {
 
     public TileEntityBasicMultiMachine(Machine<?> type) {
         super(type);
-        this.fluidHandler = type.has(FLUID) ? LazyHolder.of(() -> new BasicMultiMachineFluidHandler(this)) : LazyHolder.empty();
-        this.energyHandler = type.has(ENERGY) ? LazyHolder.of(() -> new BasicMultiMachineEnergyHandler(this, type.amps(),type.has(GENERATOR))) : LazyHolder.empty();
+        //this.fluidHandler = type.has(FLUID) ? LazyHolder.of(() -> new BasicMultiMachineFluidHandler(this)) : LazyHolder.empty();
+        //this.energyHandler = type.has(ENERGY) ? LazyHolder.of(() -> new BasicMultiMachineEnergyHandler(this, type.amps(),type.has(GENERATOR))) : LazyHolder.empty();
         // TODO
         /*
         recipeHandler.setup((tile, tag) -> new MultiMachineRecipeHandler<TileEntityMultiMachine>(tile, tag) {
